+++ frontend/src/app/App.tsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import LandingPage from '../pages/landing';
+import LoggedPage from '../pages/logged';
 import configureStore from '../store';
 
 const store = configureStore({});
 
 const App = () => (
   <Provider store={store}>
-    <LandingPage/>
+    {/* <LandingPage/> */}
+    <LoggedPage/>
   </Provider>
 );