+++ frontend/src/pages/logged/budget-container/components/budget-panes/components/header/components/add-transaction/components/AddTransactionButton.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { Button, ListGroup } from "react-bootstrap";
+
+interface Props {
+  showModal: () => void;
+}
+
+const AddTransactionButton = ({ showModal }: Props) => {
+  return (
+    <Button onClick={showModal}>
+        Manage Users
+    </Button>
+  );
+};
+
+export default AddTransactionButton;