+++ frontend/src/pages/logged/budget-container/components/budget-panes/index.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Tab } from "react-bootstrap";
+import Transaction from "./components/transaction";
+
+const BudgetPanes = () => {
+  const budgetsMock = Array(6).fill(0);
+
+  return (
+    <Tab.Content>
+      {budgetsMock.map((_, index) => (
+        <Transaction budgetItemIndex={index} key={index} />
+      ))}
+    </Tab.Content>
+  );
+};
+
+export default BudgetPanes;