+++ frontend/src/pages/logged/budget-container/components/budget-tabs/components/add/components/AddBudgetButton.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import { ListGroup } from "react-bootstrap";
+
+interface Props {
+  showModal: () => void;
+}
+
+const AddBudgetButton = ({ showModal }: Props) => {
+  return (
+    <ListGroup.Item
+      as="li"
+      className="d-flex justify-content-between align-items-start"
+      action
+    >
+      <div className="ms-2 me-auto">
+        <div className="fw-bold" onClick={showModal}>
+          Add new Budget!
+        </div>
+      </div>
+    </ListGroup.Item>
+  );
+};
+
+export default AddBudgetButton;