+++ frontend/src/pages/logged/budget-container/components/budget-tabs/index.tsx
@@ -1,0 +1,33 @@
+import React from "react";
+import { ListGroup, Pagination} from "react-bootstrap";
+import Add from "./components/add";
+import BudgetTabItem from "./components/budget-tab-item";
+import Searchbar from "./components/searchbar";
+
+const BudgetTabs = () => {
+  const budgetsMock = Array(6).fill(0);
+
+  let activePaginationTab: number = 1;
+  let paginations = [];
+
+  for (let number = 1; number <= 5; number++) {
+    paginations.push(
+      <Pagination.Item key={number} active={number === activePaginationTab}>
+        {number}
+      </Pagination.Item>,
+    );
+  }
+
+  return (
+    <ListGroup>
+      <Add/>
+      <Searchbar/>
+      {budgetsMock.map((_, index) => (
+        <BudgetTabItem budgetItemIndex={index} key={index} />
+      ))}
+      <Pagination>{paginations}</Pagination>
+    </ListGroup>
+  );
+};
+
+export default BudgetTabs;